ping2telegram/requirements.txt
pyTelegramBotAPI==3.2.0
PyYAML==3.12
